redesign module
delete duble
