--- src/form.js
@@ -1,76 +1,0 @@
-'use strict';
-
-(function() {
-  var formContainer = document.querySelector('.overlay-container');
-  var formOpenButton = document.querySelector('.reviews-controls-new');
-  var formCloseButton = document.querySelector('.review-form-close');
-
-  var reviewForm = document.querySelector('.review-form');
-  var reviewUser = reviewForm['review-name'];
-  var reviewText = reviewForm['review-text'];
-  var reviewMark = reviewForm['review-mark'];
-  var reviewGroupMark = document.querySelector('.review-form-group-mark');
-
-  var formValidation = require('./form_validation');
-  var getDateExpire = require('./form_setCookies');
-
-  var cookies = require('browser-cookies');
-  reviewUser.value = cookies.get('reviewUser');
-  reviewMark.value = cookies.get('reviewMark') || 3;
-
-  formOpenButton.addEventListener('click', function(evt) {
-    evt.preventDefault();
-    formContainer.classList.remove('invisible');
-    reviewText.required = reviewMark.value < 3;
-    formValidation();
-  });
-
-  formCloseButton.addEventListener('click', function(evt) {
-    evt.preventDefault();
-    formContainer.classList.add('invisible');
-  });
-
-  // Обработчик клика на блок с оценками
-  reviewGroupMark.addEventListener('change', function() {
-    reviewText.required = reviewMark.value < 3;
-    formValidation();
-  });
-
-  // Обработчик события нажатия клавиши
-  reviewForm.addEventListener('keyup', function() {
-    formValidation();
-  });
-
-  reviewForm.addEventListener('submit', function(evt) {
-    evt.preventDefault();
-
-    var dateToExpire = getDateExpire();
-
-    cookies.set('reviewUser', reviewUser.value, dateToExpire);
-    cookies.set('reviewMark', reviewMark.value, dateToExpire);
-    this.submit();
-  });
-
-  // //Вычисояем кол-во дней для хранения Куки
-  // //Вычесляется как:
-  // //1) Определяем кол-во дней, прошедших с последнего дня рождения
-  // //2) Прибавляем к текущей дате
-  //function getDateExpire() {
-  //  var today = new Date();
-  //  var dateBD = new Date(today.getFullYear(), 3, 8);
-  //  var dateExp;
-  //
-  //  if (dateBD > today) {
-  //    dateBD.setFullYear(dateBD.getFullYear() - 1);
-  //    dateExp = +today - +dateBD;
-  //  } else {
-  //    dateExp = +today - +dateBD;
-  //  }
-  //
-  //  var dateToExpire = +Date.now() + +dateExp;
-  //
-  //  return new Date(dateToExpire).toUTCString();
-  //}
-
-
-})();
